# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OCMS & Store LaPet Omnichannel Master API
  description: >
    Hợp đồng giao tiếp API chuẩn hóa toàn diện (Master OpenAPI Specification) giữa
    Hệ thống Quản trị Vận hành Đa kênh OCMS và Website Bán Hàng / Cổng Đại lý (store-lapet),
    đồng thời tích hợp liên thông với ERP Odoo và Kênh Zalo OA.
    
    Tài liệu này bao quát toàn bộ các nhóm tính năng vận hành:
    1. **Orders (Đơn hàng & Tiếp nhận chuẩn):** Cổng tiếp nhận đơn hàng bất đồng bộ, tra cứu trạng thái, duyệt đơn Odoo, hủy đơn và thống kê.
    2. **Products & Inventory (Sản phẩm & Tồn kho):** Danh mục sản phẩm, bảng giá đại lý/bán lẻ, kiểm tra tồn kho Odoo thời gian thực.
    3. **Promotions & Pricing Engine (Khuyến mãi & Động cơ tính giá):** Động cơ tính chiết khấu bậc thang tự động cho giỏ hàng, tư vấn ưu đãi bằng AI.
    4. **Customers & Dealers (Khách hàng & Đại lý B2B):** Xác thực OTP/Mật khẩu, hồ sơ đại lý, hạn mức công nợ, doanh số và quản lý danh bạ CRM.
    5. **Omnichannel Chat (Hội thoại & Zalo):** Quản lý luồng trò chuyện, gửi tin nhắn trao đổi, tự động bóc tách đơn hàng AI từ chat.
    6. **Webhooks & Sync (Thanh toán & Đồng bộ):** Webhook thanh toán tự động VietQR và kích hoạt đồng bộ dữ liệu Odoo ERP.
  version: 2.0.0
  contact:
    name: OCMS Integration Architecture Team
    email: dev@lapet.vn
servers:
  - url: http://localhost:3000
    description: Môi trường phát triển cục bộ (Local Development)
  - url: https://api.ocms.lapet.vn
...
paths:
...
  /api/v1/orders:
    get:
      tags:
        - Orders
      summary: Danh sách đơn hàng trong trung tâm quản trị OCMS
      description: Lấy danh sách đơn hàng đã lưu trong OCMS, hỗ trợ bộ lọc nâng cao và phân trang.
      operationId: listOrders
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
        - name: search
          in: query
          description: Tìm theo mã đơn, tên khách, số điện thoại
          schema:
            type: string
        - name: state
          in: query
          description: Trạng thái đơn hàng Odoo (draft, sale, done, cancel, pending)
          schema:
            type: string
        - name: salesperson
          in: query
          description: Lọc theo nhân viên phụ trách bán hàng
          schema:
            type: string
        - name: from
          in: query
          description: Ngày bắt đầu (YYYY-MM-DD)
          schema:
            type: string
        - name: to
          in: query
          description: Ngày kết thúc (YYYY-MM-DD)
          schema:
            type: string
      responses:
        '200':
          description: Danh sách đơn hàng phân trang
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListResponse'
  /api/v1/orders/{id}:
    get:
      tags:
        - Orders
      summary: Chi tiết đơn hàng nội bộ kèm toàn bộ dòng sản phẩm
      description: Lấy đầy đủ chi tiết đơn hàng, dòng sản phẩm, chính sách khuyến mãi đã áp dụng và file PDF báo giá Odoo.
      operationId: getOrderDetail
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID đơn hàng nội bộ OCMS (UUID) hoặc Odoo Order ID
          schema:
            type: string
      responses:
        '200':
          description: Chi tiết đơn hàng
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResponse'
        '404':
          description: Không tìm thấy đơn hàng

    put:
      tags:
        - Orders
      summary: Chỉnh sửa chi tiết đơn hàng (sản phẩm, giá, CK, ghi chú) qua Router
      description: >
        Cho phép nhân viên điều chỉnh danh sách sản phẩm, số lượng, đơn giá, chiết khấu, ghi chú đơn và ghi chú giao việc.
        Lệnh cập nhật tự động đi qua Universal Router Gateway (cổng 3001) sang đúng máy chủ Odoo đích (Test hoặc Prod).
      operationId: updateOrderDetail
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID đơn hàng nội bộ OCMS hoặc Odoo Order ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdatePayload'
      responses:
        '200':
          description: Cập nhật đơn hàng thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderUpdateResponse'
        '404':
          description: Không tìm thấy đơn hàng cần chỉnh sửa
  /api/v1/orders/{id}/activity:
    post:
      tags:
        - Orders
      summary: Thêm mới hoặc cập nhật Ghi chú giao việc (Hoạt động Odoo mail.activity) qua Router
      description: >
        Tạo mới hoặc cập nhật nội dung ghi chú giao việc cho đơn hàng.
        Hành động được tự động điều phối qua Universal Router Gateway sang đúng máy chủ Odoo đích (Test hoặc Prod).
      operationId: saveOrderActivity
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID đơn hàng nội bộ OCMS hoặc Odoo Order ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - summary
              properties:
                summary:
                  type: string
                  description: Nội dung ghi chú giao việc
                  example: "đã giao 4/9 xuat nhap PTQuyen"
      responses:
        '200':
          description: Cập nhật ghi chú giao việc thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  activitySummary:
                    type: string
                    nullable: true
                    example: "đã giao 4/9 xuat nhap PTQuyen"

    delete:
      tags:
        - Orders
      summary: Xóa Ghi chú giao việc (Hoạt động Odoo mail.activity) qua Router
      description: >
        Xóa bỏ hoạt động / ghi chú giao việc khỏi đơn hàng.
        Lệnh xóa được tự động điều phối qua Universal Router Gateway sang Odoo đích (mail.activity.unlink).
      operationId: deleteOrderActivity
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID đơn hàng nội bộ OCMS hoặc Odoo Order ID
          schema:
            type: string
      responses:
        '200':
          description: Đã xóa ghi chú giao việc thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  activitySummary:
                    type: string
                    nullable: true
                    example: null
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    OrderUpdatePayload:
      type: object
      properties:
        client_order_ref:
          type: string
          nullable: true
          description: Ghi chú đơn hàng (chỉ dẫn giao nhận / client reference)
          example: "Giao trong giờ hành chính"
        activity_summary:
          type: string
          nullable: true
          description: Ghi chú giao việc (Hoạt động Odoo mail.activity)
          example: "đã giao 4/9 xuat nhap PTQuyen"
        lines:
          type: array
          description: Danh sách chi tiết các dòng sản phẩm của đơn hàng
          items:
            $ref: '#/components/schemas/OrderUpdateLineItem'
    OrderUpdateLineItem:
      type: object
      required:
        - product_id
        - product_uom_qty
        - price_unit
      properties:
        id:
          type: integer
          nullable: true
          description: ID dòng sản phẩm Odoo (sale.order.line). Bỏ trống nếu là dòng thêm mới.
          example: 12580
        product_id:
          type: integer
          description: ID biến thể sản phẩm Odoo (product.product)
          example: 1042
        product_uom_qty:
          type: number
          minimum: 0
          description: Số lượng sản phẩm
          example: 5
        price_unit:
          type: number
          minimum: 0
          description: Đơn giá sản phẩm (VND)
          example: 25000
        discount:
          type: number
          minimum: 0
          maximum: 100
          default: 0
          description: Tỷ lệ chiết khấu theo dòng (%)
          example: 10
        name:
          type: string
          nullable: true
          description: Tên / mô tả dòng sản phẩm
          example: "Pate Thịt Cừu Cho Chó 100g"
        _deleted:
          type: boolean
          default: false
          description: Đánh dấu true nếu muốn xóa dòng sản phẩm này khỏi đơn hàng
          example: false
    OrderUpdateResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Cập nhật đơn hàng thành công"
        order:
          type: object
        router:
          type: object
          properties:
            routed:
              type: boolean
              example: true
            target:
              type: string
              example: "odoo_test"
            target_name:
              type: string
              example: "Odoo ERP (Môi trường Test)"
...
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/OrderSummaryItem'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
    OrderSummaryItem:
      type: object
      properties:
        id:
          type: string
        orderCode:
          type: string
        odooOrderId:
          type: integer
          nullable: true
        partnerName:
          type: string
        dateOrder:
          type: string
          format: date-time
        amountTotal:
          type: number
        state:
          type: string
        salesperson:
          type: string
    OrderDetailResponse:
      type: object
      properties:
        order:
          type: object
          properties:
            id:
              type: string
            orderCode:
              type: string
            odooOrderId:
              type: integer
            partnerName:
              type: string
            amountTotal:
              type: number
            lines:
              type: array
              items:
                type: object
                properties:
                  productName:
                    type: string
                  productSku:
                    type: string
                  quantity:
                    type: number
                  priceUnit:
                    type: number
                  priceSubtotal:
                    type: number
...